UI Changes 1. Added UserLoans to the Redux Store 2. Added redux-logger middleware 2. Created first version of the screen for Loan application
