+++ LoanManagement/ClientApp/package.json
@@ -13,6 +13,7 @@
     "react-router-redux": "^5.0.0-alpha.8",
     "react-scripts": "1.0.17",
     "redux": "^3.7.2",
+    "redux-logger": "^3.0.6",
     "redux-thunk": "^2.2.0",
     "rimraf": "^2.6.2"
   },
