+++ LoanManagement/ClientApp/src/App.js
@@ -4,13 +4,13 @@ import Layout from './components/Layout';
 import Home from './components/Home';
 import Counter from './components/Counter';
 import FetchData from './components/FetchData';
-import FetchUserLoans from './components/FetchUserLoans';
+import UserLoans from './components/user-loans/user-loans.component';
 
 export default () => (
   <Layout>
     <Route exact path='/' component={Home} />
     <Route path='/counter' component={Counter} />
     <Route path='/fetchdata/:startDateIndex?' component={FetchData} />
-    <Route path='/fetchUserLoans' component={FetchUserLoans} />
+    <Route path='/fetchUserLoans' component={UserLoans} />
   </Layout>
 );
